Refactor Azure upload code to a separate package
as part of this, move also the AWS upload code to the same directory
