+++ go.mod
@@ -3,8 +3,8 @@ module github.com/osbuild/osbuild-composer
 go 1.12
 
 require (
-	github.com/aws/aws-sdk-go v1.25.37
 	github.com/Azure/azure-storage-blob-go v0.8.0
+	github.com/aws/aws-sdk-go v1.25.37
 	github.com/coreos/go-systemd v0.0.0-20190719114852-fd7a80b32e1f
 	github.com/gobwas/glob v0.2.3
 	github.com/google/uuid v1.1.1
